+++ bf_email_management/models/__init__.py
@@ -10,4 +10,5 @@ from . import mail_compose_message
 from . import mail_message
 from . import mail_notification
 from . import mail_scheduled_message
+from . import res_config_settings
 from . import res_partner